+++ components/atoms/label/type.tsx
