+++ data/idol.ts
